+++ analytics/upress_reporting/models/aggregators/download_data_aggregator.py
@@ -5,6 +5,7 @@ import re
 
 from logger import createLog
 from model import Edition, Item, Link
+from models.aggregators.aggregator import Aggregator
 from models.data.interaction_event import InteractionEvent, InteractionType, UsageType
 from model.postgres.item import ITEM_LINKS
 from managers import DBManager
@@ -19,95 +20,42 @@ TIMESTAMP_REGEX = r"\[.+\]"
 REFERRER_REGEX = r"https://drb-qa.nypl.org/"
 
 
-class DownloadDataAggregator:
+class DownloadDataAggregator(Aggregator):
     """
     Parses S3 download logs and generates list of DownloadEvents, each corresponding 
     to a single download request.
     """
 
-    def __init__(self, publisher, date_range):
-        self.publisher = publisher
-        self.date_range = date_range
-
-        self.s3_client = boto3.client("s3")
+    def __init__(self, *args):
+        super().__init__(*args)
         self.bucket_name = os.environ.get("DOWNLOAD_BUCKET", None)
         self.log_path = os.environ.get("DOWNLOAD_LOG_PATH", None)
 
         self.logger = createLog("download_data_aggregator")
-        self._setup_db_manager()
+        self.setup_db_manager()
 
-    def pull_download_events(self):
+    def pull_interaction_events(self):
         '''
-        Returns list of DownloadEvents in a given reporting period.
+        Returns list of download InteractionEvents in a given reporting period.
         '''
         download_events = []
 
         for date in self.date_range:
             folder_name = date.strftime("%Y/%m/%d")
-            batch = self._load_batch(folder_name)
-            downloads_per_day = self._parse_logs(batch)
+            batch = self.load_batch(self.log_path, self.bucket_name, 
+                                    folder_name)
+            downloads_per_day = self.parse_logs_in_batch(batch, self.bucket_name)
             download_events.extend(downloads_per_day)
-
+        
         self.db_manager.closeConnection()
+
         return download_events
 
-    def _setup_db_manager(self):
-        self.db_manager = DBManager(
-            user=os.environ.get("POSTGRES_USER", None),
-            pswd=os.environ.get("POSTGRES_PSWD", None),
-            host=os.environ.get("POSTGRES_HOST", None),
-            port=os.environ.get("POSTGRES_PORT", None),
-            db=os.environ.get("POSTGRES_NAME", None),
-        )
-        self.db_manager.generateEngine()
-        self.db_manager.createSession()
-
-    def _load_batch(self, log_folder):
-        prefix = self.log_path + log_folder + "/"
-        paginator = self.s3_client.get_paginator("list_objects_v2")
-        page_iterator = paginator.paginate(
-            Bucket=self.bucket_name, Prefix=prefix)
-        return page_iterator
-
-    def _parse_logs(self, batch):
+    def match_log_info_with_drb_data(self, log_object):
         '''
-        The edition title, identifier, and timestamp are parsed out of the
-        S3 server access log files for UMP download requests.
+        Check that the S3 server access log object contains a download request. 
+        If so, we parse out the edition title, identifier, and timestamp.
         '''
-        downloads_in_batch = []
-
-        for log_file in batch:
-            if "Contents" not in log_file:
-                path = self._redact_s3_path(log_file["Prefix"])
-                raise DownloadParsingError(
-                    f"Log files in path {path} do not exist.")
-            else:
-                for content in log_file["Contents"]:
-                    curr_key = str(content["Key"])
-                    log_object_dict = self.s3_client.get_object(
-                        Bucket=self.bucket_name, Key=f"{curr_key}"
-                    )
-                    for i in log_object_dict["Body"].iter_lines():
-                        log_object_dict = i.decode("utf8")
-                        interaction_event = self._match_log_info_with_drb_data(
-                            log_object_dict)
-                        if interaction_event:
-                            downloads_in_batch.append(
-                                interaction_event)
-
-        return downloads_in_batch
-
-    def _redact_s3_path(self, path):
-        '''
-        Used to remove sensitive data from S3 prefix before passing to error message.
-        Example input = "logs/123456789/us-east-1/ump-pdf-repository/2024/1/1"
-        Example output: "logs/NYPL_AWS_ID/us-east-1/ump-pdf-repository/2024/1/1"
-        '''
-        split_path = path.split("/")
-        split_path[1] = "NYPL_AWS_ID"
-        return "/".join(split_path)
-
-    def _match_log_info_with_drb_data(self, log_object):
         matchRequest = re.search(REQUEST_REGEX, log_object)
         matchReferrer = re.search(REFERRER_REGEX, log_object)
 
@@ -186,8 +134,3 @@ class DownloadDataAggregator:
             self.logger.error(
                 "Unable to parse nypl_login flags...")
         return {}
-
-
-class DownloadParsingError(Exception):
-    def __init__(self, message=None):
-        self.message = message